+++ externals/drivers/cortexM4/lpc43xx/src/sdmmc_18xx_43xx.c
@@ -232,7 +232,7 @@ static void prv_process_csd(LPC_SDMMC_T *pSDMMC)
 		}
 	}
 
-	g_card_info->card_info.device_size = g_card_info->card_info.blocknr << 9;	/* blocknr * 512 */
+	g_card_info->card_info.device_size = (uint64_t) g_card_info->card_info.blocknr << 9;	/* blocknr * 512 */
 }
 
 /* Puts current selected card in trans state */
@@ -481,7 +481,7 @@ uint32_t Chip_SDMMC_Acquire(LPC_SDMMC_T *pSDMMC, mci_card_struct *pcardinfo)
 }
 
 /* Get the device size of SD/MMC card (after enumeration) */
-int32_t Chip_SDMMC_GetDeviceSize(LPC_SDMMC_T *pSDMMC)
+uint64_t Chip_SDMMC_GetDeviceSize(LPC_SDMMC_T *pSDMMC)
 {
 	return g_card_info->card_info.device_size;
 }
